requirements.txt
altair==5.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
commonmark==0.9.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mutagen==1.47.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==5.27.2
pyarrow==16.1.0
pycryptodomex==3.20.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==12.6.0
rpds-py==0.18.1
sentry-sdk==1.45.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
spotify_dl==8.9.0
spotipy==2.24.0
streamlit==1.36.0
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
watchdog==4.0.1
websockets==12.0
yt-dlp==2024.7.2
